Made either timestamp field completely optional
If the updatedAt or createdAt option is an empty string, the plugin will
not update that field.
This allows for schemas where maybe just createdAt is wanted.

In the query middleware I also added a check for, and remove, a possible
default value that the schema might have on an existing createdAt field.